Add endpoint to list all courses for master SU and move clone course. Fixes #287
