+++ plugins/tenant/clonecourse/package.json
@@ -1,0 +1,4 @@
+{
+  "displayName": "Copy course to tenant",
+  "version": "0.1.0"
+}
