+++ plugins/tenant/course/package.json
@@ -1,4 +1,4 @@
 {
-  "displayName": "Copy course to tenant",
+  "displayName": "List tenant courses",
   "version": "0.1.0"
 }
